requirements.txt
bs4>=4.11.1
pandas>=1.1..5
python-dotenv
selenium>=4.9.1

